Initial upload
This is the initial upload of this repository. It already includes about
1k lines of code, and several test categorical logic gate datasets and
accompanying cortex profiles that have been saved.

There are two primary classes, neuron and cortex. A neuron is the
fundamental element of a neural network, and it contains all the methods
associated with a neurons loading, firing, learning, transfer functions,
etc.

A cortex is a collection of neurons, and manages all connections between
neurons.

The test scripts provided are good places to start learning how this
stuff works.